refactor(FCMChannel): 페키지 이동
