+++ src/main/kotlin/gdsc/plantory/event/notification/WaterCycleEventListener.kt
@@ -6,7 +6,7 @@ import com.google.firebase.messaging.FirebaseMessaging
 import com.google.firebase.messaging.FirebaseMessagingException
 import com.google.firebase.messaging.Message
 import com.google.firebase.messaging.Notification
-import gdsc.plantory.fcm.domain.FCMChannel
+import gdsc.plantory.event.FCMChannel
 import org.slf4j.LoggerFactory
 import org.springframework.scheduling.annotation.Async
 import org.springframework.stereotype.Component
